+++ src/lambdaisland/witchcraft/gallery/carrier_trait.clj
@@ -5,18 +5,19 @@
             [lambdaisland.witchcraft.citizens :as c]
             [lambdaisland.witchcraft.events :as e]))
 
-(c/make-trait
- :carrier
- {:init {:inventory (wc/make-inventory {:capacity 18})}})
+(defn init! []
+  (c/make-trait
+   :carrier
+   {:init {:inventory (wc/make-inventory {:capacity 18})}})
 
-(e/listen!
- :citizens/npcright-click
- ::carrier
- (fn [e]
-   (when (c/has-trait? (:NPC e) :carrier)
-     (wc/open-inventory
-      (:clicker e)
-      (:inventory @(c/npc-trait (:NPC e) :carrier))))))
+  (e/listen!
+   :citizens/npcright-click
+   ::carrier
+   (fn [e]
+     (when (c/has-trait? (:NPC e) :carrier)
+       (wc/open-inventory
+        (:clicker e)
+        (:inventory @(c/npc-trait (:NPC e) :carrier)))))))
 
 (comment
   (def jonny (create-npc :player "jonny"))